Employment, housing, community development
I provided moving services to my local community for most of my life.  In 1998 I started using temporary labor contractors.  I was tired of dealing with payroll issues.  I would show up to pick these men up who were tired, hungry, and no disrespect to them was obvious that they had not had laundry or shower services for a while.  
In 2000 I withdrew my "Harley" fund and opened a storefront and traded a place to live to a dozen people in exchange for being available to take people to jobs, make cold calls to potential employers.  TLC was what we called our labor company whose mission was to lower the cost of living or the cost of doing business for our members by providing basic need services which included day storage, a Continental breakfast, a meal if they were to go out on assignment, advocacy, a ride to and from their job assignment, and a higher wage.

WCSC was the 501c3 which umbrellaed TLC and A Hand Up Construction whose mission was to build low-cost community infrastructure, training, and employment.  By 2004 WCSC was a partner with our local Habitat and HUD.  WCSC was designated by HUD as a CBDO, community-based development organization as such WCSC developed a 7-acre site where 14 low-income homes were built.